+++ vector/src/main/AndroidManifest.xml
@@ -98,9 +98,9 @@
                 <category android:name="android.intent.category.OPENABLE" />
             </intent-filter>
         </activity>
-        <activity android:name=".features.signout.SignedOutActivity" />
+        <activity android:name=".features.signout.hard.SignedOutActivity" />
         <activity
-            android:name=".features.signout.SoftLogoutActivity"
+            android:name=".features.signout.soft.SoftLogoutActivity"
             android:windowSoftInputMode="adjustResize" />
         <!-- Services -->
