+++ vector/src/main/java/im/vector/riotx/core/di/ScreenComponent.kt
@@ -49,7 +49,7 @@ import im.vector.riotx.features.roomdirectory.RoomDirectoryActivity
 import im.vector.riotx.features.roomdirectory.createroom.CreateRoomActivity
 import im.vector.riotx.features.settings.VectorSettingsActivity
 import im.vector.riotx.features.share.IncomingShareActivity
-import im.vector.riotx.features.signout.SoftLogoutActivity
+import im.vector.riotx.features.signout.soft.SoftLogoutActivity
 import im.vector.riotx.features.ui.UiStateRepository
 
 @Component(
