+++ vector/src/main/java/im/vector/riotx/features/signout/soft/SoftLogoutAction.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2019 New Vector Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package im.vector.riotx.features.signout.soft
+
+import im.vector.riotx.core.platform.VectorViewModelAction
+
+sealed class SoftLogoutAction : VectorViewModelAction {
+    object RetryLoginFlow : SoftLogoutAction()
+    data class PasswordChanged(val password: String) : SoftLogoutAction()
+    object TogglePassword : SoftLogoutAction()
+
+    data class SignInAgain(val password: String) : SoftLogoutAction()
+    // TODO Add reset pwd...
+}
