+++ vector/src/main/java/im/vector/riotx/features/signout/soft/epoxy/LoginHeaderItem.kt
@@ -14,15 +14,14 @@
  * limitations under the License.
  */
 
-package im.vector.riotx.features.signout
+package im.vector.riotx.features.signout.soft.epoxy
 
-import im.vector.riotx.core.platform.VectorViewModelAction
+import com.airbnb.epoxy.EpoxyModelClass
+import im.vector.riotx.R
+import im.vector.riotx.core.epoxy.VectorEpoxyHolder
+import im.vector.riotx.core.epoxy.VectorEpoxyModel
 
-sealed class SoftLogoutAction : VectorViewModelAction {
-    object RetryLoginFlow : SoftLogoutAction()
-    data class PasswordChanged(val password: String) : SoftLogoutAction()
-    object TogglePassword : SoftLogoutAction()
-
-    data class SignInAgain(val password: String) : SoftLogoutAction()
-    // TODO Add reset pwd...
+@EpoxyModelClass(layout = R.layout.item_login_header)
+abstract class LoginHeaderItem : VectorEpoxyModel<LoginHeaderItem.Holder>() {
+    class Holder : VectorEpoxyHolder()
 }
