+++ vector/src/main/java/im/vector/riotx/features/signout/soft/epoxy/LoginPasswordFormItem.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package im.vector.riotx.features.signout.epoxy
+package im.vector.riotx.features.signout.soft.epoxy
 
 import android.os.Build
 import android.text.Editable
