+++ vector/src/main/java/im/vector/riotx/features/signout/soft/epoxy/LoginRedButtonItem.kt
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package im.vector.riotx.features.signout.epoxy
+package im.vector.riotx.features.signout.soft.epoxy
 
-import android.widget.TextView
+import android.widget.Button
 import com.airbnb.epoxy.EpoxyAttribute
 import com.airbnb.epoxy.EpoxyModelClass
 import im.vector.riotx.R
@@ -24,18 +24,22 @@ import im.vector.riotx.core.epoxy.VectorEpoxyHolder
 import im.vector.riotx.core.epoxy.VectorEpoxyModel
 import im.vector.riotx.core.extensions.setTextOrHide
 
-@EpoxyModelClass(layout = R.layout.item_login_title)
-abstract class LoginTitleItem : VectorEpoxyModel<LoginTitleItem.Holder>() {
+@EpoxyModelClass(layout = R.layout.item_login_red_button)
+abstract class LoginRedButtonItem : VectorEpoxyModel<LoginRedButtonItem.Holder>() {
 
     @EpoxyAttribute var text: String? = null
+    @EpoxyAttribute var listener: (() -> Unit)? = null
 
     override fun bind(holder: Holder) {
         super.bind(holder)
 
-        holder.textView.setTextOrHide(text)
+        holder.button.setTextOrHide(text)
+        holder.button.setOnClickListener {
+            listener?.invoke()
+        }
     }
 
     class Holder : VectorEpoxyHolder() {
-        val textView by bind<TextView>(R.id.itemLoginTitleText)
+        val button by bind<Button>(R.id.itemLoginRedButton)
     }
 }
